Refactor boost content into a new component
